Fix: Gerrit folder hierarchy handling
Also clean up repository clone output to a single line per repo.
